refactor: Rearrange some properties.
